dtw/requirements.txt
Django==2.2.10
djangorestframework==3.9.4
requests==2.21.0
psycopg2-binary==2.8.2
gunicorn==19.9.0
django-debug-toolbar==1.11
django-extensions==2.1.6
coverage==4.5.3
ruamel.yaml>=0.15.34,<0.16.0
drf-yasg==1.16.0
django-enumchoicefield==1.1.0
